// src/kong_svelte/src/lib/services/canister/tcycles.d.ts
export interface CanisterSettings {
  'freezing_threshold' : [] | [bigint],
  'controllers' : [] | [Array<Principal>],
  'reserved_cycles_limit' : [] | [bigint],
  'memory_allocation' : [] | [bigint],
  'compute_allocation' : [] | [bigint],
}
export type ChangeIndexId = { 'SetTo' : Principal } |
  { 'Unset' : null };
export interface CmcCreateCanisterArgs {
  'subnet_selection' : [] | [SubnetSelection],
  'settings' : [] | [CanisterSettings],
}
export interface CreateCanisterArgs {
  'from_subaccount' : [] | [Uint8Array | number[]],
  'created_at_time' : [] | [bigint],
  'amount' : bigint,
  'creation_args' : [] | [CmcCreateCanisterArgs],
}
// ...
export type SubnetSelection = { 'Filter' : SubnetFilter } |
  { 'Subnet' : { 'subnet' : Principal } };
// ...
export declare const idlFactory: IDL.InterfaceFactory;